Make sure download URL is up-to-date
