+++ lib/src/main/java/net/dhleong/staypuft/impl/ExpansionDownloaderEngine.kt
@@ -142,7 +142,9 @@ internal class ExpansionDownloaderEngine(
             if (old != null && nameMatches && sizeMatches && !localExists
                 && old.downloaded < old.size)  {
                 // attempt to continue downloading the old file
-                old
+                // (but make sure to use the latest url)
+                Log.v("staypuft", "Continue with $old\n${old.url}\n${new.url}")
+                old.copy(url = new.url)
             } else if (old == null || !nameMatches || !sizeMatches || !localExists) {
                 // return and save the new file
                 new
